requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
